feature: homare is injected into source code!
